Explain defaultDefinitions (#5026)
* Explain defaultDefinitions

I was unclear what this `defaultDefinitions` was for, so I added an example of why it's needed. It's probably not worded very well as-is, but I at least wanted to note that the `"$ref"`s in the `schema.properties` were referring to what goes into `defaultDefinitions`.

* Moved explanation of definitions

Turns out we can actually add `definitions` to `schema`. Brilliant!