Vet Tec: pre-filled bank information is not showing on the review page#16850 (#9564)
When completing a 0994 form, a user's bank information is filled out for them on the bank information page. Currently any pre-filled information on that page is not displayed upon reaching the review page.

Log in and complete a 0994 form as a user with pre-existing bank information.
Note the account information on the Bank Information page
Continue to the Review/submit page
Expand the Personal Information section
The same bank information should be displayed here, but none is shown.